+++ src/main/java/projeto/fitosollos/repositories/CulturaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import projeto.fitosollos.entities.Cultura;
 
 @Repository
-public interface CulturaRepository extends JpaRepository<Cultura, Long> {
+public interface CulturaRepository extends JpaRepository<Cultura, Integer> {
 
 }
